requirements.txt
tensorflow-cpu
numpy
tqdm